Fix `User` routes and write `Produk` routes

+++ routes/user.js
@@ -1,10 +1,15 @@
 const router = require('express').Router();
 
-const { mustLogin } = require('../middlewares/mustLogin');
+const { mustLogin, mustBeAdmin } = require('../middlewares/must');
 
 const { User } = require('../models');
 
 router.delete('/:id', mustLogin, (req, res, next) => {
+  if (req.params.id !== req.user.id && req.user.role !== "admin") {
+    res.status(403).json({message: "If you're not doing the operation for your own account, admin role required"});
+    return;
+  }
+
   User.destroy({ where: { id: req.params.id } })
     .then((result) => {
       res.json({ result });
@@ -13,6 +18,11 @@ router.delete('/:id', mustLogin, (req, res, next) => {
 });
 
 router.put('/:id', mustLogin, (req, res, next) => {
+  if (req.params.id !== req.user.id && req.user.role !== "admin") {
+    res.status(403).json({message: "If not deleting your own account, admin role required"});
+    return;
+  }
+
   User.update(req.body, {
     where: {
       id: req.params.id
@@ -23,7 +33,7 @@ router.put('/:id', mustLogin, (req, res, next) => {
   }).catch((error) => { next(error) });
 });
 
-router.get('/:id', (req, res, next) => {
+router.get('/:id', mustLogin, mustBeAdmin, (req, res, next) => {
   User.findByPk(req.params.id, { attributes: { exclude: ["password"] } })
     .then((user) => {
       res.json({ result: user });
@@ -31,7 +41,7 @@ router.get('/:id', (req, res, next) => {
     .catch((error) => { next(error) });
 });
 
-router.get('/', (_req, res, next) => {
+router.get('/', mustLogin, mustBeAdmin, (_req, res, next) => {
   User.findAll({ attributes: { exclude: ["password"] } })
     .then((users) => {
       res.json({ result: users });
